+++ test/Reader/FeedSetTest.php
@@ -1,17 +1,16 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-feed for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-feed/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-feed/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\Feed\Reader;
+namespace LaminasTest\Feed\Reader;
 
+use Laminas\Feed\Reader\FeedSet;
 use PHPUnit_Framework_TestCase;
 use ReflectionMethod;
-use Zend\Feed\Reader\FeedSet;
 
 class FeedSetTest extends PHPUnit_Framework_TestCase
 {
@@ -30,7 +29,7 @@ class FeedSetTest extends PHPUnit_Framework_TestCase
      */
     public function testAbsolutiseUri($link, $uri)
     {
-        $method = new ReflectionMethod('Zend\Feed\Reader\FeedSet', 'absolutiseUri');
+        $method = new ReflectionMethod('Laminas\Feed\Reader\FeedSet', 'absolutiseUri');
         $method->setAccessible(true);
 
         $this->assertEquals('http://example.com/feed', $method->invoke($this->feedSet, $link, $uri));