+++ CHANGELOG.md
@@ -18,8 +18,8 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-- [#5](https://github.com/zendframework/zend-feed/pull/5) fixes the enclosure
+- [zendframework/zend-feed#5](https://github.com/zendframework/zend-feed/pull/5) fixes the enclosure
   length check to allow zero and integer strings.
-- [#2](https://github.com/zendframework/zend-feed/pull/2) ensures that the
+- [zendframework/zend-feed#2](https://github.com/zendframework/zend-feed/pull/2) ensures that the
   routine for "absolutising" a link in `Reader\FeedSet` always generates a URI
   with a scheme.